apps/ui/pyproject.toml
[project]
name = "cogstack-ui"
version = "0.1.0"
description = "CogStack Leadgen — operations console"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.2.1",
    "pydantic>=2.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cogstack_ui"]
